requirements.txt
python-dotenv>=0.21.0
langchain>=0.0.172
langchain-google-genai>=1.0.10
google-generativeai>=0.8.3
pydantic>=1.10.7
pytest>=7.4.0